# backend/src/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Get Mumm — Homemade Food Delivery API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu items and categories
  - name: chefs
    description: Home chefs and partners
  - name: blog
    description: Blog posts and recipes
  - name: testimonials
    description: Customer and office testimonials
  - name: subscriptions
    description: Subscription plans and packages
  - name: contact
    description: Contact and inquiry forms
  - name: stats
    description: Site stats and summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu/categories:
    get:
      operationId: listCategories
      tags: [menu]
      summary: List all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /menu/items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dietary
          in: query
          schema:
            type: ["string", "null"]
            enum: [vegetarian, vegan, low_cal, gluten_free, null]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu/items/featured:
    get:
      operationId: getFeaturedItems
      tags: [menu]
      summary: Get featured/popular menu items for the homepage
      responses:
        "200":
          description: Featured items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu/items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found

  /chefs:
    get:
      operationId: listChefs
      tags: [chefs]
      summary: List home chefs/partners
      responses:
        "200":
          description: List of chefs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chef"

  /chefs/{id}:
    get:
      operationId: getChef
      tags: [chefs]
      summary: Get a chef profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chef profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chef"
        "404":
          description: Not found

  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts and recipes
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [blog, recipe, null]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"

  /blog/posts/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post or recipe
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List customer and office testimonials
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [customer, office, chef, null]
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"

  /subscriptions:
    get:
      operationId: listSubscriptionPlans
      tags: [subscriptions]
      summary: List subscription plans and packages
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionPlan"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a general contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"

  /office-inquiry:
    post:
      operationId: submitOfficeInquiry
      tags: [contact]
      summary: Submit a corporate/office bulk order inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeInquiryInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"

  /stats/summary:
    get:
      operationId: getSiteSummary
      tags: [stats]
      summary: Homepage stats — meals delivered, chefs empowered, happy customers
      responses:
        "200":
          description: Site summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Category:
      type: object
      required: [id, name, nameAr, slug, description, descriptionAr, imageUrl, itemCount]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        imageUrl:
          type: string
        itemCount:
          type: integer

    MenuItem:
      type: object
      required: [id, name, nameAr, description, descriptionAr, price, categoryId, categoryName, categoryNameAr, imageUrl, dietary, isAvailable, isFeatured, chefName, chefNameAr]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryNameAr:
          type: string
        imageUrl:
          type: string
        dietary:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        chefName:
          type: string
        chefNameAr:
          type: string
        rating:
          type: ["number", "null"]
        prepTimeMinutes:
          type: ["integer", "null"]

    Chef:
      type: object
      required: [id, name, nameAr, bio, bioAr, imageUrl, specialties, specialtiesAr, itemCount, rating, joinedYear]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        bio:
          type: string
        bioAr:
          type: string
        imageUrl:
          type: string
        specialties:
          type: array
          items:
            type: string
        specialtiesAr:
          type: array
          items:
            type: string
        itemCount:
          type: integer
        rating:
          type: number
        joinedYear:
          type: integer

    BlogPost:
      type: object
      required: [id, title, titleAr, slug, excerpt, excerptAr, content, contentAr, imageUrl, type, publishedAt, author, authorAr, readTimeMinutes]
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        excerptAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        imageUrl:
          type: string
        type:
          type: string
          enum: [blog, recipe]
        publishedAt:
          type: string
        author:
          type: string
        authorAr:
          type: string
        readTimeMinutes:
          type: integer
        tags:
          type: array
          items:
            type: string

    Testimonial:
      type: object
      required: [id, name, nameAr, quote, quoteAr, type, rating, avatarUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        quote:
          type: string
        quoteAr:
          type: string
        type:
          type: string
          enum: [customer, office, chef]
        rating:
          type: integer
        avatarUrl:
          type: string
        company:
          type: ["string", "null"]
        companyAr:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        roleAr:
          type: ["string", "null"]

    SubscriptionPlan:
      type: object
      required: [id, name, nameAr, description, descriptionAr, price, period, features, featuresAr, isPopular, targetAudience]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        period:
          type: string
          enum: [weekly, monthly, daily]
        features:
          type: array
          items:
            type: string
        featuresAr:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
        targetAudience:
          type: string
          enum: [individual, family, office]

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        message:
          type: string
        subject:
          type: string

    OfficeInquiryInput:
      type: object
      required: [companyName, contactName, email, headCount, message]
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        headCount:
          type: integer
        deliveryArea:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, one_time]
        message:
          type: string

    ContactResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    SiteSummary:
      type: object
      required: [mealsDelivered, chefsEmpowered, happyCustomers, deliveryAreas, avgRating, yearsOfService]
      properties:
        mealsDelivered:
          type: integer
        chefsEmpowered:
          type: integer
        happyCustomers:
          type: integer
        deliveryAreas:
          type: integer
        avgRating:
          type: number
        yearsOfService:
          type: integer