# backend/src/api-zod/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Get Mumm — Homemade Food Delivery API
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPostType } from './blogPostType';

export interface BlogPost {
  id: number;
  title: string;
  titleAr: string;
  slug: string;
  excerpt: string;
  excerptAr: string;
  content: string;
  contentAr: string;
  imageUrl: string;
  type: BlogPostType;
  publishedAt: string;
  author: string;
  authorAr: string;
  readTimeMinutes: number;
  tags?: string[];
}