# backend/src/api-zod/generated/types/category.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Get Mumm — Homemade Food Delivery API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  description: string;
  descriptionAr: string;
  imageUrl: string;
  itemCount: number;
}