# backend/src/api-zod/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Get Mumm — Homemade Food Delivery API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  email: string;
  phone?: string;
  message: string;
  subject?: string;
}