# backend/src/api-zod/generated/types/officeInquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Get Mumm — Homemade Food Delivery API
 * OpenAPI spec version: 0.1.0
 */
import type { OfficeInquiryInputFrequency } from './officeInquiryInputFrequency';

export interface OfficeInquiryInput {
  companyName: string;
  contactName: string;
  email: string;
  phone?: string;
  headCount: number;
  deliveryArea?: string;
  frequency?: OfficeInquiryInputFrequency;
  message: string;
}