# backend/src/api-zod/generated/types/testimonial.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Get Mumm — Homemade Food Delivery API
 * OpenAPI spec version: 0.1.0
 */
import type { TestimonialType } from './testimonialType';

export interface Testimonial {
  id: number;
  name: string;
  nameAr: string;
  quote: string;
  quoteAr: string;
  type: TestimonialType;
  rating: number;
  avatarUrl: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  companyAr?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  roleAr?: string | null;
}